use cache controls when processing whole tile sets
