+++ src/mjolnir/graphoptimizer.cc
@@ -1,9 +1,23 @@
 #include "mjolnir/graphoptimizer.h"
-#include <valhalla/midgard/logging.h>
+#include "valhalla/mjolnir/graphtilebuilder.h"
 
 #include <ostream>
 #include <set>
 #include <boost/format.hpp>
+#include <sstream>
+#include <iostream>
+#include <string>
+#include <vector>
+#include <map>
+#include <utility>
+
+#include <valhalla/midgard/logging.h>
+#include <valhalla/midgard/pointll.h>
+#include <valhalla/baldr/tilehierarchy.h>
+#include <valhalla/baldr/graphid.h>
+#include <valhalla/baldr/graphconstants.h>
+#include <valhalla/baldr/graphreader.h>
+
 
 using namespace valhalla::midgard;
 using namespace valhalla::baldr;
@@ -130,6 +144,10 @@ void GraphOptimizer::Optimize(const boost::property_tree::ptree& pt) {
       if (tilebuilder.size() == 0) {
         continue;
       }
+
+      // Check if we need to clear the tile cache
+      if(graphreader_.OverCommitted())
+        graphreader_.Clear();
 
       // Copy existing header. No need to update any counts or offsets.
       GraphTileHeader existinghdr = *(tilebuilder.header());
